requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
boto==2.49.0
boto3==1.38.33
botocore==1.38.33
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
dogpile.cache==1.2.2
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.1
frozenlist==1.7.0
future==1.0.0
geoip2==5.1.0
google-auth==2.40.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maxminddb==2.7.0
mdurl==0.1.2
multidict==6.4.4
oic==1.7.0
packaging==25.0
paramiko==3.5.1
pbr==6.1.1
prometheus_client==0.22.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyjwkest==1.4.2
pymemcache==4.0.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
rucio==37.4.0
rucio-clients==37.4.0
s3transfer==0.13.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
statsd==4.0.1
stevedore==5.4.1
stomp.py==8.2.0
tabulate==0.9.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
